requirements.txt
ansicon==1.89.0 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
anyio==4.8.0 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
attrs==25.1.0 ; python_version >= "3.10" and python_version < "4.0"
awsebcli==3.21.0 ; python_version >= "3.10" and python_version < "4.0"
backports-tarfile==1.2.0 ; python_version >= "3.10" and python_version < "3.12"
blessed==1.20.0 ; python_version >= "3.10" and python_version < "4.0"
botocore==1.35.99 ; python_version >= "3.10" and python_version < "4.0"
build==1.2.2.post1 ; python_version >= "3.10" and python_version < "4.0"
cachecontrol==0.14.2 ; python_version >= "3.10" and python_version < "4.0"
cement==2.10.14 ; python_version >= "3.10" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.10" and python_version < "4.0" and (sys_platform == "linux" or sys_platform == "darwin" or platform_python_implementation == "PyPy")
charset-normalizer==3.4.1 ; python_version >= "3.10" and python_version < "4.0"
cleo==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0"
crashtest==0.4.1 ; python_version >= "3.10" and python_version < "4.0"
cryptography==44.0.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "linux"
distlib==0.3.9 ; python_version >= "3.10" and python_version < "4.0"
django-cleanup==9.0.0 ; python_version >= "3.10" and python_version < "4.0"
django-countries==7.6.1 ; python_version >= "3.10" and python_version < "4.0"
django-filter==24.3 ; python_version >= "3.10" and python_version < "4.0"
django-rest-knox==5.0.2 ; python_version >= "3.10" and python_version < "4.0"
django-seed==0.3.1 ; python_version >= "3.10" and python_version < "4.0"
django==4.2.19 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework-simplejwt==5.4.0 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.15.2 ; python_version >= "3.10" and python_version < "4.0"
drf-spectacular==0.28.0 ; python_version >= "3.10" and python_version < "4.0"
dulwich==0.22.7 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.2 ; python_version == "3.10"
faker==36.1.1 ; python_version >= "3.10" and python_version < "4.0"
fastjsonschema==2.21.1 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.17.0 ; python_version >= "3.10" and python_version < "4.0"
findpython==0.6.2 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==8.6.1 ; python_version >= "3.10" and python_version < "3.12"
inflection==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
installer==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
jaraco-classes==3.4.0 ; python_version >= "3.10" and python_version < "4.0"
jaraco-context==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
jaraco-functools==4.1.0 ; python_version >= "3.10" and python_version < "4.0"
jeepney==0.8.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "linux"
jinxed==1.3.0 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
jmespath==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2024.10.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.10" and python_version < "4.0"
keyring==25.6.0 ; python_version >= "3.10" and python_version < "4.0"
markdown==3.7 ; python_version >= "3.10" and python_version < "4.0"
more-itertools==10.6.0 ; python_version >= "3.10" and python_version < "4.0"
msgpack==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.10.1 ; python_version >= "3.10" and python_version < "4.0"
pbs-installer==2025.2.12 ; python_version >= "3.10" and python_version < "4.0"
pillow==11.1.0 ; python_version >= "3.10" and python_version < "4.0"
pkginfo==1.12.1.2 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==4.3.6 ; python_version >= "3.10" and python_version < "4.0"
poetry-core==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
poetry-plugin-export==1.9.0 ; python_version >= "3.10" and python_version < "4.0"
poetry==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
psycopg-binary==3.1.20 ; python_version >= "3.10" and python_version < "4.0" and implementation_name != "pypy"
psycopg-pool==3.2.4 ; python_version >= "3.10" and python_version < "4.0"
psycopg==3.1.20 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0" and (sys_platform == "linux" or sys_platform == "darwin" or platform_python_implementation == "PyPy")
pyjwt==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
pypiwin32==223 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
pyproject-hooks==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
pywin32-ctypes==0.2.3 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
pywin32==308 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
rapidfuzz==3.12.1 ; python_version >= "3.10" and python_version < "4.0"
referencing==0.36.2 ; python_version >= "3.10" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
rpds-py==0.22.3 ; python_version >= "3.10" and python_version < "4.0"
secretstorage==3.3.3 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "linux"
semantic-version==2.10.0 ; python_version >= "3.10" and python_version < "4.0"
setuptools==75.8.0 ; python_version >= "3.10" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.10" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.10" and python_version < "4.0"
termcolor==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.2.1 ; python_version == "3.10"
tomlkit==0.13.2 ; python_version >= "3.10" and python_version < "4.0"
toposort==1.10 ; python_version >= "3.10" and python_version < "4.0"
trove-classifiers==2025.2.18.16 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2025.1 ; python_version >= "3.10" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.20 ; python_version >= "3.10" and python_version < "4.0"
virtualenv==20.29.2 ; python_version >= "3.10" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.10" and python_version < "4.0"
xattr==1.1.4 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "darwin"
zipp==3.21.0 ; python_version >= "3.10" and python_version < "3.12"
zstandard==0.23.0 ; python_version >= "3.10" and python_version < "4.0"
